fix(settings): split seedingAllowed() out of the 'use server' dev-seed module
Next.js requires every top-level export of a 'use server' file to be an
async Server Action; seedingAllowed() is a plain sync env check, which broke
the production build outright ("Server Actions must be async functions").
Move it to dev-seed-allowed.ts (no 'use server') and have both dev-seed.ts
and the Developer page import it from there. Verified with a native
`next build` and `docker compose build core-front`, both previously failing.
