+++ core-front/apps/shell/app/settings/developer/page.tsx
@@ -1,6 +1,6 @@
 import { requireAuth } from '@/lib/session'
 import DeveloperSettings from '@/components/DeveloperSettings'
-import { seedingAllowed } from '@/lib/dev-seed'
+import { seedingAllowed } from '@/lib/dev-seed-allowed'
 
 // Settings → Developer: dev-only tools for testing the software. Auth-gated like
 // every settings page; the seed action itself is also guarded server-side via the
