+++ core-front/apps/shell/src/lib/dev-seed.ts
@@ -1,5 +1,6 @@
 'use server'
 import { ApiError, createServerApiClient } from '@eerp/core-front/server'
+import { seedingAllowed } from './dev-seed-allowed'
 
 // Settings -> Developer: populates the workspace with realistic-looking demo
 // records through the SAME generic entity API (POST /{entity}) any other write
@@ -9,21 +10,6 @@ import { ApiError, createServerApiClient } from '@eerp/core-front/server'
 // bulk, irreversible write and must never land in a real tenant. Only a UI
 // affordance exists to trigger this — no cron/API surface, so an admin has to
 // be signed in and on the Developer settings page to run it.
-
-/**
- * NODE_ENV alone can't tell "a real production deployment" apart from "a dev/local
- * deployment running the production Next build" — the standalone Docker image (see
- * core-front/Dockerfile) always sets NODE_ENV=production, since that's required for
- * `next start` to run correctly, even when compose.yml is standing it up for local
- * dev. ALLOW_DEMO_SEED lets such a deployment opt back in explicitly, the same way
- * COOKIE_SECURE overrides the NODE_ENV-derived default in session-cookies.ts.
- */
-export function seedingAllowed(): boolean {
-  if (process.env.ALLOW_DEMO_SEED !== undefined) {
-    return process.env.ALLOW_DEMO_SEED === 'true'
-  }
-  return process.env.NODE_ENV !== 'production'
-}
 
 export interface SeedEntityResult {
   entity: string
